Change the close interface to return a Future
As the socket closing is treated async, we need to ensure that we can chain
events in Future as needed to handle callback.
The error is as well thus returned as a future if we have any.

Example updated to be able to quit the process once the connection is closed.